pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools", "setuptools-scm"]

[project]
name = "brainbodyroutinebackend"
version = "1.0.0"
# ...
dependencies = [
  "alembic==1.13.1",
  "annotated-types==0.6.0",
  "anyio==4.3.0",
  "bcrypt==4.0.1",
  "black==24.4.0",
  "cffi==1.16.0",
  "click==8.1.7",
  "cryptography==42.0.6",
  "ecdsa==0.19.0",
  "exceptiongroup==1.2.1",
  "fastapi==0.110.1",
  "h11==0.14.0",
  "httptools==0.6.1",
  "idna==3.7",
  "Mako==1.3.3",
  "MarkupSafe==2.1.5",
  "mypy-extensions==1.0.0",
  "packaging==24.0",
  "passlib==1.7.4",
  "pathspec==0.12.1",
  "platformdirs==4.2.0", #   "psycopg2-binary==2.9.6",
  "psycopg2==2.9.9",
  "pyasn1==0.6.0",
  "pycparser==2.22",
  "pydantic==2.7.0",
  "pydantic_core==2.18.1",
  "python-dotenv==1.0.1",
  "python-jose==3.3.0",
  "python-multipart==0.0.9",
  "PyYAML==6.0.1",
  "rsa==4.9",
  "six==1.16.0",
  "sniffio==1.3.1",
  "SQLAlchemy==2.0.29",
  "starlette==0.37.2",
  "tomli==2.0.1",
  "typing_extensions==4.11.0",
  "uvicorn==0.29.0",
  "uvloop==0.19.0",
  "watchfiles==0.21.0",
  "websockets==12.0",
]
# ...
